refactor(status): regroup components/ into chrome/summary/table/charts/incidents
22 components (flat) become 5 focused subfolders:

  chrome/    FilterChip, FilterPills, InfoButton, KeyboardHelpOverlay,
             LightModeToggle, Popover, RefreshButton
  summary/   AvgLatencyCard, IncidentsCard, MttrCard, OverallUptimeCard,
             StatusBanner, SummaryStats
  table/     LatencySparkline, SegmentTooltip, ServiceDetailPanel,
             ServiceRow, SkeletonRow, SubServiceRow, UptimeBar (+ test)
  charts/    LatencyChart, WeekdayUptimeChart
  incidents/ IncidentCard, IncidentDetail, IncidentFilterChips, IncidentList

All imports updated:
- `$lib/components/*` in +page.svelte now goes through the subfolders
- `../types/`, `../aggregation/`, etc. inside components now `../../`
- Cross-subfolder sibling imports (summary → chrome/InfoButton,
  table/ServiceDetailPanel → charts/LatencyChart, etc.) rewritten

`git mv` preserves blame for the moved files; content stays byte-
identical to the previous commit (only import paths updated).
