+++ sites/status.arolariu.ro/src/lib/components/incidents/IncidentCard.svelte
@@ -1,6 +1,6 @@
 <script lang="ts">
-  import type {Incident} from "../types/status";
-  import {formatDuration} from "../aggregation/formatDuration";
+  import type {Incident} from "../../types/status";
+  import {formatDuration} from "../../aggregation/formatDuration";
   import IncidentDetail from "./IncidentDetail.svelte";
 
   interface Props {
